+++ src/js/index.js
@@ -432,6 +432,18 @@ class TrezorConnect {
         return await call({ method: 'stellarSignTransaction', ...params });
     }
 
+    static tezosGetAddress: $T.TezosGetAddress = async (params) => {
+        return await call({ method: 'tezosGetAddress', ...params });
+    }
+
+    static tezosGetPublicKey: $T.TezosGetPublicKey = async (params) => {
+        return await call({ method: 'tezosGetPublicKey', ...params });
+    }
+
+    static tezosSignTransaction: $T.TezosSignTransaction = async (params) => {
+        return await call({ method: 'tezosSignTransaction', ...params });
+    }
+
     static verifyMessage: $T.VerifyMessage = async (params) => {
         return await call({ method: 'verifyMessage', ...params });
     }
@@ -441,7 +453,16 @@ class TrezorConnect {
     }
 
     static dispose = (): void => {
-        // TODO
+        iframe.dispose();
+        if (_popupManager) {
+            _popupManager.close();
+        }
+    }
+
+    static cancel = (): void => {
+        if (_popupManager) {
+            _popupManager.emit(POPUP.CLOSED);
+        }
     }
 
     static renderWebUSBButton = (className: ?string): void => {