+++ src/js/popup/view/notification.js
@@ -15,16 +15,17 @@ export const showFirmwareUpdateNotification = (device: $PropertyType<UnexpectedD
     if (!device.features) return;
     const { features } = device;
     const release = getLatestRelease([ features.major_version, features.minor_version, features.patch_version ]);
+    if (!release) return;
 
     const view = views.getElementsByClassName('firmware-update-notification');
     const notification = document.createElement('div');
     notification.className = 'firmware-update-notification notification-item';
     notification.innerHTML = view.item(0).innerHTML;
 
-    if (release && release.beta) {
-        const button = notification.getElementsByClassName('notification-button')[0];
-        button.setAttribute('href', 'https://beta-wallet.trezor.io/');
-    }
+    const button = notification.getElementsByClassName('notification-button')[0];
+    const url = release.channel === 'beta' ? 'https://beta-wallet.trezor.io/' : 'https://wallet.trezor.io/';
+    const version = release.version.join('.');
+    button.setAttribute('href', `${url}?fw=${version}`);
 
     container.appendChild(notification);