+++ src/js/popup/view/passphraseOnDevice.js
@@ -3,7 +3,7 @@
 import { container, showView } from './common';
 import type { DeviceMessage } from '../../types/uiRequest';
 
-export const initPassphraseOnDeviceView = (payload: $PropertyType<DeviceMessage, 'payload'>): void => {
+export const passphraseOnDeviceView = (payload: $PropertyType<DeviceMessage, 'payload'>): void => {
     showView('passphrase-on-device');
 
     const deviceName: HTMLElement = container.getElementsByClassName('device-name')[0];